fix: make Cycle 572 and 576 audit sources complete
