+++ outputs/physical_dynamical_metric_source_law_bridge_tournament_cycle576_receipt_2026_07_22.json
@@ -4,20 +4,22 @@
   "claim_type_author_hint": "bounded_theorem",
   "authority": "none",
   "audit": "unset",
-  "runner_sha256": "40a6507717669dc388d8b916ec6f4d95c6643bf677a88e767e9abed554981359",
-  "note_sha256": "10eb871032f2517d8710db0d3b43c9344cfa89fabd4b3ff216fdde51250b8248",
+  "runner_sha256": "371e4f65878109afd95dcb57670312dfe8666fabde811603f0f47da379235f3c",
+  "regge_support_sha256": "1211c739167236636f35a362f374a0204de8bb8476812a00cb4d3f5d0fa75430",
+  "plaquette_support_sha256": "652801a920d65735e0156ec9ad1958604cdc3b297af5c33a957e4b9b8896d440",
+  "note_sha256": "cc17666541e86b536069c86883557e3574414c7b3324aaaac5840a2094c2fcd7",
   "tests_passed": 10,
   "tests_total": 10,
   "pass": true,
   "clean_clone_provenance": {
     "repair_base": "origin/main@98b0cfa3b1b5e1977dce76471638a8094f58e964",
     "canonical_source_note": "docs/FINITE_REGGE_PLAQUETTE_SCATTERING_DIAGNOSTICS_CYCLE576_BOUNDED_THEOREM_NOTE_2026-07-22.md",
-    "pinned_dependencies": 7,
+    "pinned_dependencies": 9,
     "missing_paths": [],
     "untracked_dependencies": [],
-    "Cycle572_receipt_sha256": "bc3decba5fad793a177cf16823a03e7f786d717f1f68a269f7fb0859e089a675",
-    "Cycle572_runner_sha256": "8d9d2f53b672847fb6678ebbb6654699361f4afadf0a43daad620d99db96cdc5",
-    "Cycle572_note_sha256": "46fc6d404d5b7ef973d1ef6475c802575eaedb17122b9c425217d837382acd4e"
+    "Cycle572_receipt_sha256": "312ea93982e59dca5385a47d3054994c2d9921fa9e5ed042ddb8bc6d082b5d87",
+    "Cycle572_runner_sha256": "ee3cae3d548dbf7f28c848cdb2710521483aa66e46b4e60363627b7d2d36e9b0",
+    "Cycle572_note_sha256": "72e37ba7301e2aa6926c666304440bf370e2e22158d69e26262440d51bf07ecc"
   },
   "supplied_parameters": {
     "regge_update_scale": 0.025,
