+++ scripts/physical_dynamical_metric_source_law_bridge_tournament_cycle576_2026_07_22.py
@@ -14,9 +14,8 @@ called Einstein gravity.
 
 from __future__ import annotations
 
-from functools import lru_cache
 from hashlib import sha256
-from itertools import permutations, product
+from itertools import product
 import json
 import math
 from pathlib import Path
@@ -25,14 +24,14 @@ import sys
 from time import perf_counter
 
 import numpy as np
-from scipy.linalg import expm
-from scipy.sparse.linalg import expm_multiply
 
 
 ROOT = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT / "scripts"))
 
 import frontier_cubic_coxeter_regge_second_variation_3plus1_2026_06_09 as regge
+import physical_dynamical_metric_source_law_bridge_cycle576_regge_support_2026_07_22 as regge_support
+import physical_dynamical_metric_source_law_bridge_cycle576_plaquette_support_2026_07_22 as plaquette_support
 
 
 NOTE = ROOT / (
@@ -58,6 +57,8 @@ FAIL = 0
 AUDIT_TIMEOUT_SEC = 300
 AUDIT_INPUT_PATHS = (
     "docs/FINITE_REGGE_PLAQUETTE_SCATTERING_DIAGNOSTICS_CYCLE576_BOUNDED_THEOREM_NOTE_2026-07-22.md",
+    "scripts/physical_dynamical_metric_source_law_bridge_cycle576_regge_support_2026_07_22.py",
+    "scripts/physical_dynamical_metric_source_law_bridge_cycle576_plaquette_support_2026_07_22.py",
     "outputs/physical_source_insertion_selection_backreaction_tournament_cycle572_receipt_2026_07_22.json",
     "scripts/physical_source_insertion_selection_backreaction_tournament_cycle572_2026_07_22.py",
     "docs/FINITE_SOURCE_INSERTION_ALGEBRA_CARRIER_LABEL_SUPPORT_CYCLE572_BOUNDED_THEOREM_NOTE_2026-07-22.md",
@@ -69,12 +70,16 @@ AUDIT_INPUT_PATHS = (
 
 
 DEPENDENCIES = {
+    "scripts/physical_dynamical_metric_source_law_bridge_cycle576_regge_support_2026_07_22.py":
+        "1211c739167236636f35a362f374a0204de8bb8476812a00cb4d3f5d0fa75430",
+    "scripts/physical_dynamical_metric_source_law_bridge_cycle576_plaquette_support_2026_07_22.py":
+        "652801a920d65735e0156ec9ad1958604cdc3b297af5c33a957e4b9b8896d440",
     "outputs/physical_source_insertion_selection_backreaction_tournament_cycle572_receipt_2026_07_22.json":
-        "bc3decba5fad793a177cf16823a03e7f786d717f1f68a269f7fb0859e089a675",
+        "312ea93982e59dca5385a47d3054994c2d9921fa9e5ed042ddb8bc6d082b5d87",
     "scripts/physical_source_insertion_selection_backreaction_tournament_cycle572_2026_07_22.py":
-        "8d9d2f53b672847fb6678ebbb6654699361f4afadf0a43daad620d99db96cdc5",
+        "ee3cae3d548dbf7f28c848cdb2710521483aa66e46b4e60363627b7d2d36e9b0",
     "docs/FINITE_SOURCE_INSERTION_ALGEBRA_CARRIER_LABEL_SUPPORT_CYCLE572_BOUNDED_THEOREM_NOTE_2026-07-22.md":
-        "46fc6d404d5b7ef973d1ef6475c802575eaedb17122b9c425217d837382acd4e",
+        "72e37ba7301e2aa6926c666304440bf370e2e22158d69e26262440d51bf07ecc",
     "scripts/r3_regge_linearization_lambda1_healthy_graviton_2026_06_08.py":
         "cd70b8d2d2deb0bd539c0d33db8254205e0112356a943a046aab4c0e1ca43264",
     "docs/R3_GEOMETRIC_REGGE_LINEARIZATION_GIVES_HEALTHY_LAMBDA1_GRAVITON_NARROW_THEOREM_NOTE_2026-06-08.md":
@@ -107,11 +112,84 @@ def dependency_controls() -> dict:
         # The campaign freezes only committed dependency surfaces. The parent
         # rechecks this with git; this local field makes the contract explicit.
         tracked[name] = (ROOT / name).exists()
+    closure = source_dependency_closure()
     return {
         "expected": DEPENDENCIES,
         "observed": observed,
         "all_paths_exist": all(tracked.values()),
-        "pass": observed == DEPENDENCIES and all(tracked.values()),
+        "source_dependency_closure": closure,
+        "pass": observed == DEPENDENCIES and all(tracked.values()) and closure["pass"],
+    }
+
+
+def source_dependency_closure() -> dict:
+    source_modules = (
+        "scripts/physical_dynamical_metric_source_law_bridge_tournament_cycle576_2026_07_22.py",
+        "scripts/physical_dynamical_metric_source_law_bridge_cycle576_regge_support_2026_07_22.py",
+        "scripts/physical_dynamical_metric_source_law_bridge_cycle576_plaquette_support_2026_07_22.py",
+        "scripts/frontier_cubic_coxeter_regge_second_variation_3plus1_2026_06_09.py",
+    )
+    helper_paths = source_modules[1:]
+    ordinary_helper_imports = (
+        "import frontier_cubic_coxeter_regge_second_variation_3plus1_2026_06_09 as regge",
+        "import physical_dynamical_metric_source_law_bridge_cycle576_regge_support_2026_07_22 as regge_support",
+        "import physical_dynamical_metric_source_law_bridge_cycle576_plaquette_support_2026_07_22 as plaquette_support",
+    )
+    texts = {
+        path: (ROOT / path).read_text(encoding="utf-8")
+        for path in source_modules
+    }
+    imports = {
+        path: tuple(
+            line.strip()
+            for line in text.splitlines()
+            if line.startswith("import ") or line.startswith("from ")
+        )
+        for path, text in texts.items()
+    }
+    forbidden = (
+        "git" + " show",
+        "merge" + "-base",
+        "sub" + "process",
+        "docs/work_" + "history",
+    )
+    forbidden_counts = {
+        path: {
+            fragment: text.lower().count(fragment.lower())
+            for fragment in forbidden
+        }
+        for path, text in texts.items()
+    }
+    observed_ordinary_imports = tuple(
+        line
+        for line in imports[source_modules[0]]
+        if line in ordinary_helper_imports
+    )
+    return {
+        "source_modules": source_modules,
+        "source_module_characters": {
+            path: len(text) for path, text in texts.items()
+        },
+        "source_module_sha256": {
+            path: sha256(text.encode("utf-8")).hexdigest()
+            for path, text in texts.items()
+        },
+        "imports": imports,
+        "ordinary_helper_imports": observed_ordinary_imports,
+        "audit_input_helpers": tuple(
+            path for path in helper_paths if path in AUDIT_INPUT_PATHS
+        ),
+        "forbidden_reference_counts": forbidden_counts,
+        "pass": (
+            all(len(text) < 40_000 for text in texts.values())
+            and observed_ordinary_imports == ordinary_helper_imports
+            and all(path in AUDIT_INPUT_PATHS for path in helper_paths)
+            and not any(
+                count
+                for counts in forbidden_counts.values()
+                for count in counts.values()
+            )
+        ),
     }
 
 
@@ -152,483 +230,14 @@ def note_contract() -> dict:
     return {"required": required, "missing": missing, "pass": not missing}
 
 
-def proper_cubic_frames() -> tuple[np.ndarray, ...]:
-    frames = []
-    for order in permutations(range(3)):
-        permutation = np.zeros((3, 3), dtype=int)
-        for row, column in enumerate(order):
-            permutation[row, column] = 1
-        for signs in product((-1, 1), repeat=3):
-            frame = np.diag(signs) @ permutation
-            if round(np.linalg.det(frame)) == 1:
-                frames.append(frame)
-    frames.sort(key=lambda item: tuple(item.reshape(-1)))
-    return tuple(frames)
+FRAMES = regge_support.FRAMES
+FRAME_LOOKUP = regge_support.FRAME_LOOKUP
 
 
-FRAMES = proper_cubic_frames()
-FRAME_LOOKUP = {tuple(frame.reshape(-1)): index for index, frame in enumerate(FRAMES)}
 
+FACE_DIRECTIONS = plaquette_support.FACE_DIRECTIONS
+FACE_REPS = plaquette_support.FACE_REPS
 
-def lift_frame(frame: np.ndarray) -> np.ndarray:
-    result = np.eye(4, dtype=int)
-    result[:3, :3] = frame
-    return result
-
-
-def metric_matrix(values: np.ndarray) -> np.ndarray:
-    result = np.zeros((4, 4), dtype=complex)
-    for value, (left, right) in zip(values, regge.HCOMPS):
-        result[left, right] = value
-        result[right, left] = value
-    return result
-
-
-def metric_vector(matrix: np.ndarray) -> np.ndarray:
-    return np.asarray([matrix[left, right] for left, right in regge.HCOMPS])
-
-
-def metric_representation(frame: np.ndarray) -> np.ndarray:
-    lifted = lift_frame(frame)
-    representation = np.zeros((10, 10), dtype=float)
-    for column in range(10):
-        basis = np.zeros(10)
-        basis[column] = 1
-        representation[:, column] = metric_vector(
-            lifted @ metric_matrix(basis) @ lifted.T
-        ).real
-    return representation
-
-
-METRIC_REPS = tuple(metric_representation(frame) for frame in FRAMES)
-LIFTED_FRAMES = tuple(lift_frame(frame) for frame in FRAMES)
-
-
-def momentum_key(momentum: np.ndarray) -> tuple[float, ...]:
-    return tuple(float(round(value, 13)) for value in momentum)
-
-
-@lru_cache(maxsize=None)
-def base_edge_hessian_cached(key: tuple[float, ...]) -> np.ndarray:
-    return regge.bloch_Q(np.asarray(key, dtype=float))
-
-
-@lru_cache(maxsize=None)
-def base_deficit_source_cached(key: tuple[float, ...]) -> np.ndarray:
-    momentum = np.asarray(key, dtype=float)
-    source = np.zeros(15, dtype=complex)
-    for triangle in regge.TRI_CLASSES:
-        source += regge.tri_rows(triangle, momentum)[1]
-    return source
-
-
-def base_edge_hessian(momentum: np.ndarray) -> np.ndarray:
-    return base_edge_hessian_cached(momentum_key(momentum)).copy()
-
-
-def base_deficit_source(momentum: np.ndarray) -> np.ndarray:
-    return base_deficit_source_cached(momentum_key(momentum)).copy()
-
-
-def base_metric_hessian(momentum: np.ndarray) -> np.ndarray:
-    metric_map = regge.metric_map(momentum)
-    result = metric_map.conj().T @ base_edge_hessian(momentum) @ metric_map
-    return (result + result.conj().T) / 2
-
-
-def frame_averaged_metric_hessian(momentum: np.ndarray) -> np.ndarray:
-    result = np.zeros((10, 10), dtype=complex)
-    for representation, frame in zip(METRIC_REPS, LIFTED_FRAMES):
-        result += representation.T @ base_metric_hessian(frame @ momentum) @ representation
-    return result / len(FRAMES)
-
-
-def base_metric_source_row(momentum: np.ndarray) -> np.ndarray:
-    # Fourier row convention: delta(k)=d(k)e(k), e(k)=M(k)h(k).
-    return base_deficit_source(momentum) @ regge.metric_map(momentum)
-
-
-def frame_averaged_source_row(momentum: np.ndarray) -> np.ndarray:
-    result = np.zeros(10, dtype=complex)
-    for representation, frame in zip(METRIC_REPS, LIFTED_FRAMES):
-        result += base_metric_source_row(frame @ momentum) @ representation
-    return result / len(FRAMES)
-
-
-def continuum_gauge_metric(momentum: np.ndarray) -> np.ndarray:
-    result = np.zeros((10, 4), dtype=complex)
-    for direction in range(4):
-        tensor = np.zeros((4, 4), dtype=complex)
-        for left in range(4):
-            tensor[left, direction] += 1j * momentum[left]
-            tensor[direction, left] += 1j * momentum[left]
-        result[:, direction] = metric_vector(tensor)
-    return result
-
-
-def frame_sector_permutation(frame: np.ndarray) -> np.ndarray:
-    """Active rotation: old sector F moves to F G^{-1}."""
-    inverse = frame.T
-    dimension = 1 + 24 * 15
-    result = np.zeros((dimension, dimension), dtype=float)
-    result[0, 0] = 1
-    for old, sector in enumerate(FRAMES):
-        target = FRAME_LOOKUP[tuple((sector @ inverse).reshape(-1))]
-        source_slice = slice(1 + 15 * old, 1 + 15 * (old + 1))
-        target_slice = slice(1 + 15 * target, 1 + 15 * (target + 1))
-        result[target_slice, source_slice] = np.eye(15)
-    return result
-
-
-FRAME_SECTOR_REPS = tuple(frame_sector_permutation(frame) for frame in FRAMES)
-
-
-def frame_sector_hamiltonian(
-    momentum: np.ndarray,
-    *,
-    source_amplitude: float = 1.0,
-    include_regge: bool = True,
-    include_source: bool = True,
-) -> np.ndarray:
-    dimension = 1 + 24 * 15
-    result = np.zeros((dimension, dimension), dtype=complex)
-    sources = []
-    for index, frame in enumerate(LIFTED_FRAMES):
-        local_momentum = frame @ momentum
-        block = slice(1 + 15 * index, 1 + 15 * (index + 1))
-        if include_regge:
-            result[block, block] = REGGE_UPDATE_SCALE * base_edge_hessian(local_momentum)
-        sources.append(base_deficit_source(local_momentum))
-    joined = np.concatenate(sources)
-    if include_source:
-        result[0, 1:] = SOURCE_COUPLING * source_amplitude * joined
-        result[1:, 0] = np.conj(result[0, 1:])
-    return result
-
-
-def route_a_regge() -> dict:
-    directions = (
-        np.asarray((1, 0, 0, 0), float),
-        np.asarray((1, 1, 0, 0), float) / math.sqrt(2),
-        np.asarray((0, 0, 0, 1), float),
-        np.asarray((1, 0, 0, 1), float) / math.sqrt(2),
-        np.asarray((1, 1, 1, 1), float) / 2,
-    )
-    small = 1.0e-3
-    comparisons = []
-    for direction in directions:
-        momentum = small * direction
-        candidate = frame_averaged_metric_hessian(momentum)
-        target = regge.einstein_pairing_4d(momentum)
-        coefficient = float(np.vdot(target, candidate).real / np.vdot(target, target).real)
-        residual = float(np.linalg.norm(candidate - coefficient * target) / np.linalg.norm(candidate))
-        comparisons.append({
-            "direction": direction.tolist(),
-            "best_fit_coefficient": coefficient,
-            "relative_residual": residual,
-        })
-
-    generic = np.asarray((0.17, 0.11, 0.07, 0.13))
-    candidate = frame_averaged_metric_hessian(generic)
-    gauge = continuum_gauge_metric(generic)
-    bianchi = float(np.max(abs(candidate @ gauge)))
-    source_row = frame_averaged_source_row(generic)
-    source_ward = float(np.max(abs(source_row @ gauge)))
-    edge_bianchi = 0.0
-    deficit_ward = 0.0
-    for frame in LIFTED_FRAMES:
-        local_momentum = frame @ generic
-        edge_bianchi = max(
-            edge_bianchi,
-            float(np.max(abs(base_edge_hessian(local_momentum) @ regge.gauge_map(local_momentum)))),
-        )
-        for triangle in regge.TRI_CLASSES:
-            deficit = regge.tri_rows(triangle, local_momentum)[1]
-            deficit_ward = max(deficit_ward, float(np.max(abs(deficit @ regge.gauge_map(local_momentum)))))
-
-    covariance = 0.0
-    source_covariance = 0.0
-    for representation, frame in zip(METRIC_REPS, LIFTED_FRAMES):
-        covariance = max(
-            covariance,
-            float(np.linalg.norm(
-                representation.T @ frame_averaged_metric_hessian(frame @ generic) @ representation
-                - candidate
-            )),
-        )
-        source_covariance = max(
-            source_covariance,
-            float(np.linalg.norm(frame_averaged_source_row(frame @ generic) @ representation - source_row)),
-        )
-
-    product_residual = 0.0
-    sector_product_residual = 0.0
-    products = 0
-    for left_index, left in enumerate(FRAMES):
-        for right_index, right in enumerate(FRAMES):
-            target_index = FRAME_LOOKUP[tuple((left @ right).reshape(-1))]
-            product_residual = max(
-                product_residual,
-                float(np.linalg.norm(METRIC_REPS[left_index] @ METRIC_REPS[right_index] - METRIC_REPS[target_index])),
-            )
-            sector_product_residual = max(
-                sector_product_residual,
-                float(np.linalg.norm(
-                    FRAME_SECTOR_REPS[left_index] @ FRAME_SECTOR_REPS[right_index]
-                    - FRAME_SECTOR_REPS[target_index]
-                )),
-            )
-            products += 1
-
-    finite_hamiltonian = frame_sector_hamiltonian(generic)
-    declared_raw_source_generator = SOURCE_COUPLING * np.concatenate([
-        base_deficit_source(frame @ generic) for frame in LIFTED_FRAMES
-    ])
-    raw_source_generator_residual = float(np.linalg.norm(
-        finite_hamiltonian[0, 1:] - declared_raw_source_generator
-    ))
-    hermiticity = float(np.linalg.norm(finite_hamiltonian - finite_hamiltonian.conj().T))
-    initial = np.zeros(len(finite_hamiltonian), dtype=complex)
-    initial[0] = 1
-    evolved = expm_multiply(-1j * UPDATE_PARAMETER * finite_hamiltonian, initial)
-    restored = expm_multiply(+1j * UPDATE_PARAMETER * finite_hamiltonian, evolved)
-    deleted_source = expm_multiply(
-        -1j * UPDATE_PARAMETER * frame_sector_hamiltonian(generic, include_source=False), initial
-    )
-    deleted_response = expm_multiply(
-        -1j * UPDATE_PARAMETER * frame_sector_hamiltonian(generic, include_regge=False), initial
-    )
-    wrong_sign = expm_multiply(
-        -1j * UPDATE_PARAMETER * frame_sector_hamiltonian(generic, source_amplitude=-1.0), initial
-    )
-    zero_source = deleted_source
-    amplitudes = []
-    size_fixtures = (
-        ("TRAIN_L3", 0.6, False, 3, np.asarray((2 * np.pi / 3, 0, 0, 0))),
-        ("HELD_L4_LOW", 0.37, True, 4, np.asarray((np.pi / 2, np.pi / 2, 0, 0))),
-        ("HELD_L4_SIGN", -0.81, True, 4, np.asarray((np.pi / 2, 0, np.pi / 2, np.pi / 2))),
-    )
-    for label, amplitude, held, length, momentum in size_fixtures:
-        hamiltonian = frame_sector_hamiltonian(momentum, source_amplitude=amplitude)
-        output = expm_multiply(-1j * UPDATE_PARAMETER * hamiltonian, initial)
-        back = expm_multiply(+1j * UPDATE_PARAMETER * hamiltonian, output)
-        amplitudes.append({
-            "fixture": label,
-            "source_amplitude": amplitude,
-            "held": held,
-            "periodic_spatial_length": length,
-            "bloch_momentum": momentum.tolist(),
-            "metric_carrier_norm": float(np.linalg.norm(output[1:])),
-            "inverse_residual": float(np.linalg.norm(back - initial)),
-            "parameters_refit": 0,
-        })
-
-    h_rotated = frame_sector_hamiltonian(LIFTED_FRAMES[1] @ generic)
-    representation = FRAME_SECTOR_REPS[1]
-    update_covariance_one_frame = float(np.linalg.norm(
-        representation.T @ expm(-1j * UPDATE_PARAMETER * h_rotated) @ representation
-        - expm(-1j * UPDATE_PARAMETER * finite_hamiltonian)
-    ))
-
-    single = base_metric_hessian(generic)
-    anisotropic_control = 0.0
-    for representation_metric, frame in zip(METRIC_REPS, LIFTED_FRAMES):
-        anisotropic_control = max(
-            anisotropic_control,
-            float(np.linalg.norm(
-                representation_metric.T @ base_metric_hessian(frame @ generic) @ representation_metric - single
-            )),
-        )
-
-    return {
-        "route": "A_supplied_24_sector_actual_Regge_edge_carrier",
-        "frame_sector_status": (
-            "24 frame sectors are coordinates of a supplied finite one-excitation model; uniform coherent "
-            "frame-sector preparation/readout is supplied and not selected"
-        ),
-        "frame_average_status": (
-            "compile-time target comparison and first-order uniform-sector projection only; not a stochastic mixture "
-            "and not a separately enacted update"
-        ),
-        "source_generator": "H_source=+lambda sum_x q_x sum_local_hinges delta_hinge",
-        "source_fourier_normalization": "raw finite-range deficit symbol; no momentum-dependent normalization",
-        "momentum_dependent_source_normalization_used": False,
-        "raw_local_source_generator_residual": raw_source_generator_residual,
-        "source_called_physical_stress_energy_or_gravity": False,
-        "R3_EH_target_algebra_comparisons": comparisons,
-        "mean_best_fit_coefficient": float(np.mean([row["best_fit_coefficient"] for row in comparisons])),
-        "coefficient_spread": float(np.ptp([row["best_fit_coefficient"] for row in comparisons])),
-        "maximum_R3_relative_residual": max(row["relative_residual"] for row in comparisons),
-        "metric_Bianchi_residual": bianchi,
-        "edge_Regge_Bianchi_residual": edge_bianchi,
-        "local_deficit_source_Ward_residual": deficit_ward,
-        "frame_averaged_source_Ward_residual": source_ward,
-        "maximum_all24_metric_generator_covariance_residual": covariance,
-        "maximum_all24_source_covariance_residual": source_covariance,
-        "all576_metric_representation_products": products,
-        "all576_metric_representation_residual": product_residual,
-        "all576_sector_representation_residual": sector_product_residual,
-        "finite_edge_source_Hermiticity_residual": hermiticity,
-        "finite_state_update_inverse_residual": float(np.linalg.norm(restored - initial)),
-        "finite_state_update_norm_residual": abs(float(np.vdot(evolved, evolved).real) - 1.0),
-        "source_deletion_residual": float(np.linalg.norm(evolved - deleted_source)),
-        "metric_response_deletion_residual": float(np.linalg.norm(evolved - deleted_response)),
-        "zero_source_metric_carrier_norm": float(np.linalg.norm(zero_source[1:])),
-        "wrong_sign_response_sum_residual": float(np.linalg.norm(evolved[1:] + wrong_sign[1:])),
-        "wrong_sign_is_equally_unitary_and_not_selected": True,
-        "single_frame_anisotropic_control_residual": anisotropic_control,
-        "one_frame_finite_update_covariance_residual": update_covariance_one_frame,
-        "source_amplitude_rows": amplitudes,
-        "finite_coordinate_count": 361,
-        "finite_coordinate_layout": "one source coordinate plus 24x15 edge coordinates",
-        "local_generator_support": "one path 4-cell and its bounded hinge star; tick offsets are internal layer roles",
-        "finite_time_state_law_exact_unitary": True,
-        "bounded_depth_finite_time_Regge_circuit_compiled": False,
-        "target_equation_used_as_update": False,
-        "Einstein_equation_or_physical_gravity_derived": False,
-        "physical_site_encoding_or_intertwiner_executed": False,
-    }
-
-
-AXIAL_DIRECTIONS = np.asarray(
-    ((1, 0, 0), (-1, 0, 0), (0, 1, 0), (0, -1, 0), (0, 0, 1), (0, 0, -1)),
-    dtype=int,
-)
-FACE_DIRECTIONS = np.asarray(
-    sorted(direction for direction in product((-1, 0, 1), repeat=3) if sum(value != 0 for value in direction) == 2),
-    dtype=int,
-)
-
-
-def direction_permutation(directions: np.ndarray, frame: np.ndarray) -> np.ndarray:
-    count = len(directions)
-    result = np.zeros((count, count), dtype=float)
-    for source, vector in enumerate(directions):
-        target = int(np.where(np.all(directions == frame @ vector, axis=1))[0][0])
-        result[target, source] = 1
-    return result
-
-
-AXIAL_REPS = tuple(direction_permutation(AXIAL_DIRECTIONS, frame) for frame in FRAMES)
-FACE_REPS = tuple(direction_permutation(FACE_DIRECTIONS, frame) for frame in FRAMES)
-
-
-def plaquette_hamiltonian(
-    *,
-    gamma: float = PLAQUETTE_GAMMA,
-    eta: float = PLAQUETTE_ETA,
-    omega: float = PLAQUETTE_OMEGA,
-    anisotropic: bool = False,
-) -> np.ndarray:
-    # 6 matter rails + 12 curvature rails + 12 conjugate rails + 1 reservoir.
-    result = np.zeros((31, 31), dtype=complex)
-    weights = np.zeros((6, 12))
-    for matter, axis in enumerate(AXIAL_DIRECTIONS):
-        for face, diagonal in enumerate(FACE_DIRECTIONS):
-            weights[matter, face] = (axis @ diagonal) ** 2 - 2 / 3
-    if anisotropic:
-        weights[:, 1:] = 0
-    result[:6, 6:18] = gamma * weights / math.sqrt(12)
-    result[6:18, :6] = result[:6, 6:18].T
-    result[6:18, 18:30] = omega * np.eye(12)
-    result[18:30, 6:18] = omega * np.eye(12)
-    scalar = np.ones(12) / math.sqrt(12)
-    result[30, 6:18] = eta * scalar
-    result[6:18, 30] = eta * scalar
-    return result
-
-
-def plaquette_representation(frame_index: int) -> np.ndarray:
-    result = np.zeros((31, 31), dtype=float)
-    result[:6, :6] = AXIAL_REPS[frame_index]
-    result[6:18, 6:18] = FACE_REPS[frame_index]
-    result[18:30, 18:30] = FACE_REPS[frame_index]
-    result[30, 30] = 1
-    return result
-
-
-PLAQUETTE_REPS = tuple(plaquette_representation(index) for index in range(24))
-
-
-def route_b_plaquette(receipt: dict) -> dict:
-    hamiltonian = plaquette_hamiltonian()
-    unitary = expm(-1j * UPDATE_PARAMETER * hamiltonian)
-    initial = np.zeros(31, dtype=complex)
-    initial[30] = 1 / math.sqrt(2)
-    initial[0] = 1 / math.sqrt(2)
-    evolved = unitary @ initial
-    restored = unitary.conj().T @ evolved
-    deleted_source = expm(-1j * UPDATE_PARAMETER * plaquette_hamiltonian(eta=0.0)) @ initial
-    deleted_curvature = expm(-1j * UPDATE_PARAMETER * plaquette_hamiltonian(gamma=0.0)) @ initial
-    deleted_conjugate = expm(-1j * UPDATE_PARAMETER * plaquette_hamiltonian(omega=0.0)) @ initial
-    empty_source_initial = np.zeros(31, dtype=complex)
-    empty_source = unitary @ empty_source_initial
-    reservoir_deleted_matter_initial = np.zeros(31, dtype=complex)
-    reservoir_deleted_matter_initial[0] = 1
-    reservoir_deleted_matter = unitary @ reservoir_deleted_matter_initial
-    wrong_sign = expm(-1j * UPDATE_PARAMETER * plaquette_hamiltonian(gamma=-PLAQUETTE_GAMMA)) @ initial
-    anisotropic_h = plaquette_hamiltonian(anisotropic=True)
-
-    covariance = 0.0
-    anisotropic_covariance = 0.0
-    product_residual = 0.0
-    products = 0
-    for left_index, left in enumerate(FRAMES):
-        representation = PLAQUETTE_REPS[left_index]
-        covariance = max(covariance, float(np.linalg.norm(representation @ hamiltonian - hamiltonian @ representation)))
-        anisotropic_covariance = max(
-            anisotropic_covariance,
-            float(np.linalg.norm(representation @ anisotropic_h - anisotropic_h @ representation)),
-        )
-        for right_index, right in enumerate(FRAMES):
-            target = FRAME_LOOKUP[tuple((left @ right).reshape(-1))]
-            product_residual = max(
-                product_residual,
-                float(np.linalg.norm(PLAQUETTE_REPS[left_index] @ PLAQUETTE_REPS[right_index] - PLAQUETTE_REPS[target])),
-            )
-            products += 1
-
-    spatial_components = (0, 1, 2, 4, 5, 7)
-    momentum = np.asarray((1.0e-3, 0, 0, 0))
-    target = regge.einstein_pairing_4d(momentum)[np.ix_(spatial_components, spatial_components)]
-    candidate = PLAQUETTE_OMEGA * np.eye(6)
-    coefficient = float(np.vdot(target, candidate).real / np.vdot(target, target).real)
-    r3_residual = float(np.linalg.norm(candidate - coefficient * target) / np.linalg.norm(candidate))
-    gauge_sample = np.asarray((1, 0, 0, 0, 0, 0), dtype=float)
-
-    weights = hamiltonian[:6, 6:18]
-    return {
-        "route": "B_local_plaquette_curvature_conjugate_reciprocal_gate",
-        "local_dimension": 31,
-        "finite_coordinate_count": 31,
-        "one_excitation_resource_conserved": True,
-        "Hermiticity_residual": float(np.linalg.norm(hamiltonian - hamiltonian.conj().T)),
-        "unitarity_residual": float(np.linalg.norm(unitary.conj().T @ unitary - np.eye(31))),
-        "inverse_residual": float(np.linalg.norm(restored - initial)),
-        "source_deletion_residual": float(np.linalg.norm(evolved - deleted_source)),
-        "curvature_response_deletion_residual": float(np.linalg.norm(evolved - deleted_curvature)),
-        "conjugate_response_deletion_residual": float(np.linalg.norm(evolved - deleted_conjugate)),
-        "empty_source_carrier_response_norm": float(np.linalg.norm(empty_source[6:30])),
-        "reservoir_deleted_matter_response_norm": float(np.linalg.norm(reservoir_deleted_matter[6:30])),
-        "wrong_sign_response_difference": float(np.linalg.norm(evolved - wrong_sign)),
-        "wrong_sign_is_equally_unitary_and_not_selected": True,
-        "matter_curvature_reciprocity_residual": float(np.linalg.norm(weights - hamiltonian[6:18, :6].T)),
-        "matter_to_curvature_response_norm": float(np.linalg.norm(weights)),
-        "curvature_to_matter_backreaction_norm": float(np.linalg.norm(hamiltonian[6:18, :6])),
-        "maximum_all24_generator_covariance_residual": covariance,
-        "all576_frame_products": products,
-        "all576_frame_product_residual": product_residual,
-        "anisotropic_control_covariance_residual": anisotropic_covariance,
-        "R3_candidate": "momentum-independent six-component oscillator metric block",
-        "R3_best_fit_coefficient": coefficient,
-        "R3_relative_residual": r3_residual,
-        "R3_gauge_sample_response_norm": float(np.linalg.norm(candidate @ gauge_sample)),
-        "R3_target_match_closed": False,
-        "placement": "finite local model only; no compiled placement in a physical-site schedule",
-        "Cycle572_gamma_reused_but_not_derived": receipt["route_B_plaquette_curvature_response"]["gamma_supplied"],
-        "called_physical_stress_energy_or_gravity": False,
-        "physical_site_encoding_or_intertwiner_executed": False,
-    }
 
 
 GROVER_COIN = 2 * np.ones((12, 12), dtype=complex) / 12 - np.eye(12, dtype=complex)
@@ -986,8 +595,8 @@ def main() -> int:
     dependencies = dependency_controls()
     receipt = cycle572_receipt()
     note = note_contract()
-    route_a = route_a_regge()
-    route_b = route_b_plaquette(receipt)
+    route_a = regge_support.route_a_regge()
+    route_b = plaquette_support.route_b_plaquette(receipt)
     route_c = route_c_scattering()
     supplied = inventory()
